+++ src/components/AdminWeb/Pages/Mapas/components/MapHeader/components/CreateMapButton.tsx
@@ -2,9 +2,11 @@ import { Button, ButtonProps } from '@nextui-org/react';
 
 function CreateMapButton({ ...props }: ButtonProps) {
   return (
-    <Button {...props} size="lg" color="primary">
-      CREAR MAPA
-    </Button>
+    <>
+      <Button {...props} size="lg" className="md:flex hidden" color="primary">
+        CREAR MAPA
+      </Button>
+    </>
   );
 }
