+++ src/components/AdminWeb/Pages/Mapas/components/MapMain/components/MapCard.tsx
@@ -25,7 +25,7 @@ function MapCard({
       className={`hover:border-8 overflow-hidden transition-all hover:border-[#9F9F9F] relative flex flex-col justify-center bg-[#222222] items-center bg-gradient-to-b  w-48 h-48 rounded-3xl`}
     >
       {estaSeleccionado && (
-        <div className="flex justify-center items-center absolute w-7 h-7 rounded-full bg-white border-3 border-green-400 text-green-400 z-50 top-2 right-2">
+        <div className="flex justify-center items-center absolute w-7 h-7 rounded-full bg-white border-3 border-green-400 text-green-400 z-40 top-2 right-2">
           <svg
             xmlns="http://www.w3.org/2000/svg"
             fill="none"
@@ -43,8 +43,10 @@ function MapCard({
         </div>
       )}
       <div className="p-2 flex flex-col justify-center items-center h-full w-full text-white">
-        <span className="text-3xl">{name}</span>
-        <span className="text-sm">{originalName.slice(0, 18)}...</span>
+        <span className="text-3xl break-all">{name}</span>
+        <span className="text-sm break-all">
+          {originalName.slice(0, 18)}...
+        </span>
       </div>
       <div className="absolute opacity-[10%] h-full w-full">
         <Image
